新增工作台功能，支持查询今日营业数据、订单管理数据、菜品总览和套餐总览

+++ sky-server/src/main/java/com/sky/mapper/UserMapper.java
@@ -19,8 +19,8 @@ public interface UserMapper {
     User getById(Long userId);
 
     @Select("select count(*) from user where create_time < #{localEndTime}")
-    Double searchUserByEndTime(LocalDateTime localEndTime);
+    Integer searchUserByEndTime(LocalDateTime localEndTime);
 
     @Select("select count(*) from user where create_time between #{localBeginTime} and #{localEndTime}")
-    Double searchUserBycreateTime(LocalDateTime localBeginTime, LocalDateTime localEndTime);
+    Integer searchUserBycreateTime(LocalDateTime localBeginTime, LocalDateTime localEndTime);
 }
